Migrate lib code to TypeScript
